Improve caching strategy for InformationUpdater
Before, the caching strategy used arrays to know in memory which entries
were already in the DB. Now, it instantiates hashes whose keys are the
main business logic identifiers, and whose values are the db id's.

The benchmark was done using the dmdm method, so all
records were already stored in the DB (after the first rehearsal).
The development DB is reset before each benchmark run.

The result was as follow:
                                 user     system      total        real
cache main keys as arrays:   1.521793   0.153611   1.675404 (  1.683041)
cache CPF as hash:           1.141121   0.105883   1.247004 (  1.253813)
cache main keys as hashes:   0.307638   0.007245   0.314883 (  0.318226)