Delegate logic to methods and make code DRYer
Before, the conversion of CSV file to database entries were done by a "single
script", that is, a code that would run in order, line by line. It worked
alright but was more difficult to test.

Now the the code is separated in more methods. There's some loss of performance,
as indicated by the benchmark below, but an increase in readability.

About the benchmarking results: on the first run, all records are created on the
DB, and on the second, all records are take to the cache. So the results below
compare the two codes ("single script" and "broken in methods") when run on the
same position (first or second).

### result 1:
### run benchmark with "single script" first and collect only its result
### then run benchmark again with "broken in methods" first and collect only its result
#                      user      system     total       real
# single script:      0.714137   0.275389   0.989526   (1.157263)
# broken in methods:  0.740409   0.283729   1.024138   (1.200663)

### result 2:
### run benchmark with "broken in methods" first and collect only the result for "single script"
### then run benchmark again with "single script" first and collect only the result for "broken in methods"
#                      user      system     total       real
# single script:      0.337468   0.005192   0.342660   (0.344364)
# broken in methods:  0.351834   0.006811   0.358645   (0.362020)